requirements.txt
praw>=7.7.1
python-dotenv>=1.0.0
google-generativeai>=0.3.0
click>=8.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0